Add gymnax sampler, update sampler path
